FourDigits.cs
Write a program that takes as input a four-digit number in format abcd (e.g. 2011) and performs the following:

Calculates the sum of the digits (in our example 2 + 0 + 1 + 1 = 4) and prints it on the console.
Prints on the console the number in reversed order: dcba (in our example 1102) and prints the reversed number.
Puts the last digit in the first position: dabc (in our example 1201) and prints the result.
Exchanges the second and the third digits: acbd (in our example 2101) and prints the result.
Input

The input will consist of a single four-digit integer number on a single line.
Output

Output the result of each action on a separate line, in the same order as they are explained above:
meaning the sum comes on the first line, the reversed number on the second, and so on.